+++ Sources/VCore/Services and Managers/Device Orientation Observer/DeviceOrientationObserver.swift
@@ -35,7 +35,7 @@ import Combine
         subscription = NotificationCenter.default
             .publisher(for: UIDevice.orientationDidChangeNotification)
             .compactMap { _ in Self.getDeviceOrientation() }
-            .assign(to: \.deviceOrientation, on: self)
+            .assignWeak(to: \.deviceOrientation, on: self)
     }
 
     private static func getDeviceOrientation() -> UIDeviceOrientation {